Add rotor description to comments section
A minimum update focused on the academic side of this project. Since the rotors are a complex part of this machine, a better description was included. This in hope of further developing this module so that it can be used as a base for creating new and unique enigma machines with custom rotors, plug boards, rotations, reflections, etc.