Add rotors class to Enigma
In order to advance in the class rework, I see the necessity to create a separate class (for now) of the rotors. This is the first step towards a more modular and flexible enigma class that allows for easy enigma machine creations. Next step is to use this rotors as alphabets for simple substitution implied in the machine´s logic.